refactor: move FieldError
